Adding env. factors for natural commodity availability
Introducing the notion that environmental factors can impact the availability of natural commodities. For example, a planet's geology will affect the availability of certain ores (this will be down to the planet's geological history, core temperature, etc.). Another example would be the planet's atmosphere and climate, affecting the ability to grow certain crops of certain level of quality. Depending on the level of choice a user has in located their civilisations, this could include factoring in seasons, altitude, latitude, etc.